pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "alembic>=1.16.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "sqlalchemy[asyncio]>=2.0.41",
    "uvicorn>=0.34.2",
    "psycopg-binary>=3.2.9",
    "psycopg>=3.2.9",
    "python-dateutil>=2.9.0.post0",
    "aiosqlite>=0.21.0",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pylint>=3.3.7",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "ruff>=0.11.11",
]

[tool.mypy]
disable_error_code = ["import-untyped"]

[tool.pylint.format]
max-line-length = 88

[tool.pylint.'MESSAGES CONTROL']
disable = 'missing-function-docstring,missing-class-docstring,missing-module-docstring,too-few-public-methods'

[tool.pylint."DESIGN"]
max-args = 7
max-positional-arguments = 7
